stream-device: Allow the guest to send some information about the stream
This will allow the guest to tell its idea of numeric ID.
Using only virtual devices (like QXL or VGA) there's usually a
one to one mapping between guest and server numbering.
That is first graphic device for the guest is first display
channel for the server. This is not more true if you have device
assignment devices (pass through) as SPICE don't knows much about
these devices.
Also is possible that the guest decides to use only device assigned
devices. This causes some problems, mostly with client mouses which
requires the server to send events with guest IDs.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>